requirements.txt
SQLAlchemy
dj-database-url
django-postgrespool
gunicorn
psycopg2
whitenoise

# ********************************
# ******* django-dash core *******
# ********************************

https://github.com/barseghyanartur/django-dash/archive/master.zip
django-tinymce>=3.2.0
ipdb
ipython
radar>=0.3,<1.0
selenium
simple-timer>=0.2
Django>=3.1,<4.0
django-admin-tools>=0.8.0
django-autoslug>=1.9.8
django-debug-toolbar>=2.1
django-registration>=3.1.1
django-widget-tweaks>=1.4.8
django-template-debug==0.3.5
easy-thumbnails>=2.7.0
#lorem-ipsum-generator==0.3
https://github.com/Riamse/lorem-ipsum-generator/archive/f2e7d7aa07afb41eb27f668b0a674c409931d243.zip
